[OCA] update from github

+++ queue_job/models/queue_job.py
@@ -1,5 +1,5 @@
 # Copyright 2013-2016 Camptocamp SA
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html)
 
 import logging
 from datetime import datetime, timedelta
@@ -233,15 +233,20 @@ class QueueJob(models.Model):
 
     @api.model
     def autovacuum(self):
-        """Delete all jobs done since more than ``_removal_interval`` days.
+        """Delete all jobs done based on the removal interval defined on the
+           channel
 
         Called from a cron.
         """
-        deadline = datetime.now() - timedelta(days=self._removal_interval)
-        jobs = self.search(
-            [('date_done', '<=', deadline)],
-        )
-        jobs.unlink()
+        for channel in self.env['queue.job.channel'].search([]):
+            deadline = datetime.now() - timedelta(
+                days=int(channel.removal_interval))
+            jobs = self.search(
+                [('date_done', '<=', deadline),
+                 ('channel', '=', channel.complete_name)],
+            )
+            if jobs:
+                jobs.unlink()
         return True
 
     @api.multi
@@ -329,6 +334,9 @@ class JobChannel(models.Model):
     job_function_ids = fields.One2many(comodel_name='queue.job.function',
                                        inverse_name='channel_id',
                                        string='Job Functions')
+    removal_interval = fields.Integer(
+        default=lambda self: self.env['queue.job']._removal_interval,
+        required=True)
 
     _sql_constraints = [
         ('name_uniq',
